finsihed moving and renaming specs. Config files also moved. Started to trim helper files. Committing before merging chirag's changes to recordedit specs that removed the re-implementation of ermrestJS functionality in the test suite.
