+++ test/e2e/specs/all-features-confirmation/recordedit/add.conf.js
