+++ test/e2e/specs/all-features-confirmation/recordedit/delete.spec.js
