+++ test/e2e/specs/all-features-confirmation/recordedit/edit-delete.conf.js
