+++ test/e2e/specs/all-features/permissions-visibility.conf.js
