+++ test/e2e/specs/all-features/record/copy-btn.conf.js
