+++ test/e2e/specs/all-features/record/permissions-visibility.spec.js
