+++ test/e2e/specs/all-features/recordedit/permissions-visibility.spec.js
