+++ test/e2e/specs/all-features/recordset/permissions-visibility.spec.js
