+++ test/e2e/specs/delete-prohibited/record/no-delete-btn.conf.js
