+++ test/e2e/specs/delete-prohibited/recordedit/delete-btn.conf.js
