--- test/e2e/specs/login/login.spec.js
@@ -1,5 +1,0 @@
-describe('Chaise login app', function() {
-  it('should pass', function() {
-    expect(1).toEqual(1);
-  });
-});
