--- test/e2e/specs/viewer/helpers.js
@@ -1,5 +1,0 @@
-var chaisePage = require('../../utils/chaise.page.js');
-
-exports.testPresentation = function (tableParams) {
-
-};
